Fixed compiling most of the code on the Windows platform with Visual Studio

+++ src/renderergl1/tr_flares.cpp
@@ -362,7 +362,7 @@ void RB_RenderFlare( flare_t *f ) {
  * The coefficient flareCoeff will determine the falloff speed with increasing distance.
  */
 
-	factor = distance + size * sqrt(flareCoeff);
+	factor = distance + size * (float)sqrt((float)flareCoeff);
 	
 	intensity = flareCoeff * size * size / (factor * factor);
